// apps/admin-shadcn/src/types/openapi.d.ts
export interface components {
    schemas: {
// ...
        SessionItemDto: {
            /** @example ses_01HXYZ */
            id: string;
            /**
             * Format: date-time
             * @example 2026-01-01T00:00:00.000Z
             */
            createdAt: string;
            /**
             * Format: date-time
             * @example 2026-06-01T00:00:00.000Z
             */
            expiresAt: string;
            /** @example 127.0.0.1 */
            ipAddress: string | null;
            /** @example Mozilla/5.0 */
            userAgent: string | null;
            /** @example false */
            isCurrent: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}